Remove possibility of null property in models

+++ ShopOfThings/Business/Models/ProductFilterSearchModel.cs
@@ -3,8 +3,8 @@
     public class ProductFilterSearchModel
     {
         public Guid? StorageTypeId { set; get; }
-        public int? MinPrice { set; get; }
-        public int? MaxPrice { set; get; }
+        public decimal? MinPrice { set; get; }
+        public decimal? MaxPrice { set; get; }
         public Guid? ProductCategoryId { get; set; }
 
     }
